go back to menue if user typed 'y'

+++ main.py
@@ -5,6 +5,7 @@ import matplotlib.pyplot as plt
 import sqlite3
 from datetime import datetime
 
+
 class Evaluator:
     """ EVALUATOR ANALYZE YOUR DATA """
     db_file = "database.db"
@@ -32,8 +33,9 @@ class Evaluator:
         elif response == "2":
             self.delete_task()
         elif response == "3":
-            task = input("TYPE YOUR TASK TITLE: ")
-            self.add_task(task)
+            task = input(f"TYPE YOUR TASK TITLE: {Styles.warning}(Enter b to back){Styles.end_part}")
+            if task != "b":
+                self.add_task(task)
         elif response == "4":
             self.show_last_results(10)
         elif response == "5":
@@ -48,7 +50,6 @@ class Evaluator:
             print(f"{Styles.warning}Your response does not match any item; so choose again!{Styles.end_part}")
             self.display_menu()
 
-
     def create_tasks_table(self):
         try:
             connection = sqlite3.connect(self.db_file)
@@ -60,7 +61,6 @@ class Evaluator:
             print(f"{Styles.danger}Can't create tasks table{Styles.end_part}")
             return False
 
-
     def insert_task(self, task):
         connection = sqlite3.connect(self.db_file)
         cursor = connection.cursor()
@@ -68,7 +68,6 @@ class Evaluator:
         cursor.execute(query)
         connection.commit()
         connection.close()
-
 
     def create_results_table(self):
         connection = sqlite3.connect(self.db_file)
@@ -82,7 +81,6 @@ class Evaluator:
             print(f"{Styles.danger}Can't create tasks table{Styles.end_part}")
             return False
 
-
     def insert_result(self, result, date):
         connection = sqlite3.connect(self.db_file)
         cursor = connection.cursor()
@@ -91,7 +89,6 @@ class Evaluator:
         connection.commit()
         connection.close()
 
-
     def clean_table(self, table):
         connection = sqlite3.connect(self.db_file)
         cursor = connection.cursor()
@@ -99,7 +96,6 @@ class Evaluator:
         cursor.execute(query)
         connection.commit()
         connection.close()
-
 
     def show_tasks(self, limit=None):
         """ RETURNS ALL TASKS """
@@ -119,7 +115,6 @@ class Evaluator:
             return False
         return True
 
-
     def get_tasks(self, limit=None):
         """ RETURNS ALL TASKS """
         connection = sqlite3.connect(self.db_file)
@@ -130,7 +125,6 @@ class Evaluator:
         tasks: List[Any] = connection.execute(query).fetchall()
         connection.close()
         return tasks
-
 
     def do_we_have_result(self):
         """ returns True or False """
@@ -143,7 +137,6 @@ class Evaluator:
                 return True
             else:
                 return False
-
 
     def show_last_results(self, limit):
         """RETURNS A GRAPHICAL CHART OF LAST RESULTS"""
@@ -171,7 +164,6 @@ class Evaluator:
         else:
             print(f"{Styles.warning}There is no result to display!{Styles.end_part}")
 
-
     def get_tasks_count(self):
         """ RETURNS COUNT OF ALL OF THE TASKS """
         conn = sqlite3.connect(self.db_file)
@@ -180,33 +172,33 @@ class Evaluator:
         conn.close()
         return count
 
-
     def add_today_results(self):
         """ this method will add toda's analyze results """
-        answers = []
-        rank = 0
-        percentage_per_task = 100 / self.tasks_count
-        tasks = self.get_tasks()
-        for task_id, name in tasks:
-            task = name
-            answer = input("Did you " + task + "? " + "(y or n) ")
-            answers.append(answer)
-            if answer == "y":
-                rank += percentage_per_task
-            elif answer != "y" and answer != "n":
-                while answer != "y" and answer != "n":
-                    print(f"{Styles.warning}please answer with y or n!{Styles.end_part}")
-                    answer = input(f"{Styles.info}Did you {task} ? (y or n) {Styles.end_part}")
-        rank = round(rank, 2)
-        print(f"{Styles.success}You've done {str(rank)} % of your tasks!{Styles.end_part}")
-        if rank > 80:
-            print(f"{Styles.success}Well done! you did great!{Styles.end_part}")
-        else:
-            print(f"{Styles.warning}OOPS! Your rank is not good! tomorrow try more!{Styles.end_part}")
-        # insert result into results table
-        date = datetime.now().strftime("%Y/%m/%d")
-        self.insert_result(rank, date)
-
+        user_answer = input(f"Do you wanna coninue? {Styles.warning}(y or n){Styles.end_part} ")
+        if user_answer == "y":
+            answers = []
+            rank = 0
+            percentage_per_task = 100 / self.tasks_count
+            tasks = self.get_tasks()
+            for task_id, name in tasks:
+                task = name
+                answer = input("Did you " + task + "? " + "(y or n) ")
+                answers.append(answer)
+                if answer == "y":
+                    rank += percentage_per_task
+                elif answer != "y" and answer != "n":
+                    while answer != "y" and answer != "n":
+                        print(f"{Styles.warning}please answer with y or n!{Styles.end_part}")
+                        answer = input(f"{Styles.info}Did you {task} ? (y or n) {Styles.end_part}")
+            rank = round(rank, 2)
+            print(f"{Styles.success}You've done {str(rank)} % of your tasks!{Styles.end_part}")
+            if rank > 80:
+                print(f"{Styles.success}Well done! you did great!{Styles.end_part}")
+            else:
+                print(f"{Styles.warning}OOPS! Your rank is not good! tomorrow try more!{Styles.end_part}")
+            # insert result into results table
+            date = datetime.now().strftime("%Y/%m/%d")
+            self.insert_result(rank, date)
 
     def add_task(self, task):
         """ ADD A SINGLE TASK """
@@ -223,38 +215,43 @@ class Evaluator:
             connection.close()
         return True
 
-
     def delete_task(self):
         """ DELETE A SINGLE TASK BASED ON ID """
         # Show tasks and their ids
         there_is_task = self.show_tasks()
         if there_is_task:
-            task_id = input("TYPE TASK ID TO REMOVE: ")
-            connection = sqlite3.connect(self.db_file)
-            cursor = connection.cursor()
-            query = f'DELETE FROM tasks WHERE id="{task_id}"'
-            try:
-                cursor.execute(query)
-                connection.commit()
-                print(f"{Styles.success}Task delete successfully!{Styles.end_part}")
-            except:
-                print(f"{Styles.danger}Can not delete task with id {task_id}{Styles.end_part}")
-            finally:
-                connection.close()
-
+            task_id = input(f"TYPE TASK ID TO REMOVE: {Styles.warning}(Enter b to back){Styles.end_part}")
+            if task_id != "b":
+                connection = sqlite3.connect(self.db_file)
+                cursor = connection.cursor()
+                query = f'DELETE FROM tasks WHERE id="{task_id}"'
+                try:
+                    cursor.execute(query)
+                    connection.commit()
+                    print(f"{Styles.success}Task delete successfully!{Styles.end_part}")
+                except:
+                    print(f"{Styles.danger}Can not delete task with id {task_id}{Styles.end_part}")
+                finally:
+                    connection.close()
+            else:
+                self.display_menu()
 
     def delete_last_result(self):
         """ Delete last result """
         if self.do_we_have_result():
-            connection = sqlite3.connect(self.db_file)
-            cursor = connection.cursor()
-            query = "delete from results where id= (select id from results order by id desc limit 1);"
-            try:
-                cursor.execute(query)
-                connection.commit()
-                print(f"{Styles.success} Last result deleted successfully!{Styles.end_part}")
-            except Exception as error:
-                print(f"{Styles.danger} Sorry; can't delete last result. {error}{Styles.end_part}")
+            answer = input(f"{Styles.info}Are you sure? {Styles.end_part}{Styles.warning}(y or n){Styles.end_part}")
+            if answer == "y":
+                connection = sqlite3.connect(self.db_file)
+                cursor = connection.cursor()
+                query = "delete from results where id= (select id from results order by id desc limit 1);"
+                try:
+                    cursor.execute(query)
+                    connection.commit()
+                    print(f"{Styles.success} Last result deleted successfully!{Styles.end_part}")
+                except Exception as error:
+                    print(f"{Styles.danger} Sorry; can't delete last result. {error}{Styles.end_part}")
+            else:
+                return True
         else:
             print(f"{Styles.warning} There is no result to delete!{Styles.end_part}")
 
@@ -276,7 +273,7 @@ if __name__ == "__main__":
     # IF NOT EXISTS
     app.create_tasks_table()
     app.create_results_table()
-    
+
     continue_app = True
     plt.rcdefaults()
     while continue_app:
